Move player feature to shared directory
