+++ app/(features)/page/__tests__/PagePage.test.tsx
@@ -1,7 +1,7 @@
 import { render, act } from '@testing-library/react';
 import { SettingsProvider } from '@/app/providers/SettingsContext';
 import { BookmarkProvider } from '@/app/providers/BookmarkContext';
-import { AudioProvider } from '@/app/(features)/player/context/AudioContext';
+import { AudioProvider } from '@/app/shared/player/context/AudioContext';
 import { SidebarProvider } from '@/app/providers/SidebarContext';
 import { ThemeProvider } from '@/app/providers/ThemeContext';
 import QuranPage from '@/app/(features)/page/[pageId]/page';
