+++ app/(features)/surah/[surahId]/components/Verse.tsx
@@ -1,7 +1,7 @@
 // app/(features)/surah/[surahId]/components/Verse.tsx
 import { memo, useCallback } from 'react';
 import { Verse as VerseType, Translation } from '@/types';
-import { useAudio } from '@/app/(features)/player/context/AudioContext';
+import { useAudio } from '@/app/shared/player/context/AudioContext';
 import { useSettings } from '@/app/providers/SettingsContext';
 import { useBookmarks } from '@/app/providers/BookmarkContext';
 import { sanitizeHtml } from '@/lib/text/sanitizeHtml';
