+++ app/shared/player/components/ReciterPanel.tsx
