+++ app/shared/player/components/RepeatPanel.tsx
