+++ app/shared/player/components/Timeline.tsx
