+++ app/shared/player/components/VolumeControl.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import * as Slider from '@radix-ui/react-slider';
 import { Volume2, VolumeX } from 'lucide-react';
-import { useAudio } from '@/app/(features)/player/context/AudioContext';
+import { useAudio } from '@/app/shared/player/context/AudioContext';
 
 interface Props {
   theme: 'light' | 'dark';
