+++ app/shared/player/hooks/usePlaybackCompletion.ts
